% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{add_uncertainty}
\alias{add_uncertainty}
\title{Add uncertainty}
\usage{
add_uncertainty(
  uncertainty,
  date_symptoms,
  id_biter,
  id_case,
  use_known_source,
  buffer = 7,
  max_tries = 100
)
}
\description{
Add uncertainty
}
\keyword{internal}
