% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{boot_trees}
\alias{boot_trees}
\title{Wrapper function for bootstrapped trees: reproducibly & in parallel}
\usage{
boot_trees(
  id_case,
  id_biter,
  y_coord,
  x_coord,
  owned,
  date_symptoms,
  days_uncertain,
  lineages = NULL,
  all_chains_sequenced = FALSE,
  exclude_progen = FALSE,
  use_known_source = FALSE,
  prune = TRUE,
  si_fun,
  dist_fun,
  cutoff = 0.95,
  params,
  min_time = 1e-06,
  N = 1,
  seed = 1245,
  exp_funs = NULL,
  exp_pkgs = c("data.table", "treerabid"),
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{id_case}{id of case}

\item{id_biter}{id of biter (i.e. known biter from contact tracing data, if
unknown should be 0, if no contact tracing data, can pass NULL)}

\item{y_coord}{y coordinate of case (should be in UTM: to do = use haversine distance for long/lat)}

\item{x_coord}{x coordinate of case (should be in UTM: to do = use haversine distance for long/lat)}

\item{owned}{whether animal is owned or not (per Katie, to account for uncertainty in case locations)}

\item{date_symptoms}{case date (i.e. date symptoms started)}

\item{days_uncertain}{uncertainty in days around case date}

\item{lineages}{a data table with two columns, id_case and lineage, designating a lineage
assignment for each case, defaults to NULL which means trees wont be resolved to a phylogeny}

\item{all_chains_sequenced}{whether all chains should have atleast one sequenced case (i.e. if lineage
assignments exist), default is FALSE which mean some chains may not be linked to any sequenced case}

\item{exclude_progen}{boolean of length id_case or 1, if TRUE then case should be excluded as a potential
progenitor (i.e. if including livestock cases or other species that are dead-end transmissions)}

\item{use_known_source}{whether to assign known progenitors from contact tracing data}

\item{prune}{whether to prune links (i.e. based on Cori et al./Mancy et al.)
at a certain cutoff probability. This also results in assignment of incursions}

\item{si_fun}{the function to get the probability of a given temporal
difference between cases in days (i.e. the serial interval)}

\item{dist_fun}{the function to get the probability of a given spatial
difference between cases in meters (i.e. the dispersal kernel)}

\item{cutoff}{the probability level at which to prune links (i.e. cases can not be
linked if the distance or time difference is greater than the this \%ile of the distribution)}

\item{params}{list of parameters to pass to si_fun and dist_fun}

\item{min_time}{if uncertainty results in negative time difference between
known case pairs, set it to this value (better way to deal with propagating
uncertainty given known case pairs?)}

\item{N}{number of trees to build}

\item{seed}{seed to pass to doRNG to make trees reproducible}

\item{exp_funs}{functions to export to foreach loop (i.e. for customized
si_fun/dist_fun which relies on an external scripts), may be needed
for certain types of cluster configs}

\item{exp_pkgs}{packages to export to foreach loop, defaults to
data.table and treerabid, if other dependencies for si_fun or
dist_fun, then pass here. May be needed for
certain types of cluster configs,}
}
\value{
a data.table with bootstrapped trees
}
\description{
Wrapper function for bootstrapped trees: reproducibly & in parallel
}
