% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{build_all_links}
\alias{build_all_links}
\title{Summarize links between cases across bootstrapped transmission trees}
\usage{
build_all_links(ttrees, N)
}
\arguments{
\item{ttrees}{bootstrapped trees from \code{boot_trees}}

\item{N}{number of bootstrapped trees}
}
\value{
a data.table with all linked cases across trees by their frequency and
probability
}
\description{
Summarize links between cases across bootstrapped transmission trees
}
