% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{build_known_tree}
\alias{build_known_tree}
\title{Reconstruct known tree (i.e. from tracing data)}
\usage{
build_known_tree(id_case, id_biter, y_coord, x_coord, owned, date_symptoms)
}
\arguments{
\item{id_case}{id of case}

\item{id_biter}{id of biter (i.e. known biter from contact tracing data, if
unknown should be 0, if no contact tracing data, can pass NULL)}

\item{y_coord}{y coordinate of case (should be in UTM: to do = use haversine distance for long/lat)}

\item{x_coord}{x coordinate of case (should be in UTM: to do = use haversine distance for long/lat)}

\item{owned}{whether animal is owned or not (per Katie, to account for uncertainty in case locations)}

\item{date_symptoms}{case date (i.e. date symptoms started)}
}
\value{
a data.table with the known contact tracing tree
}
\description{
Reconstruct known tree (i.e. from tracing data)
}
\keyword{internal}
