% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{build_tree}
\alias{build_tree}
\title{Reconstruct transmission trees from a case line list.}
\usage{
build_tree(
  id_case,
  id_biter,
  y_coord,
  x_coord,
  owned,
  date_symptoms,
  days_uncertain,
  use_known_source = FALSE,
  prune = TRUE,
  si_fun,
  dist_fun,
  cutoff = 0.95,
  params
)
}
\arguments{
\item{id_case}{id of case}

\item{id_biter}{id of biter (i.e. known biter from contact tracing data, if
unknown should be 0, if no contact tracing data, can pass NULL)}

\item{y_coord}{y coordinate of case (should be in UTM: to do = use haversine distance for long/lat)}

\item{x_coord}{x coordinate of case (should be in UTM: to do = use haversine distance for long/lat)}

\item{owned}{whether animal is owned or not (per Katie, to account for uncertainty in case locations)}

\item{date_symptoms}{case date (i.e. date symptoms started)}

\item{days_uncertain}{uncertainty in days around case date}

\item{use_known_source}{whether to assign known progenitors from contact tracing data}

\item{prune}{whether to prune links (i.e. based on Cori et al./Mancy et al.)
at a certain cutoff probability. This also results in assignment of incursions}

\item{si_fun}{the function to get the probability of a given temporal
difference between cases in days (i.e. the serial interval)}

\item{dist_fun}{the function to get the probability of a given spatial
difference between cases in meters (i.e. the dispersal kernel)}

\item{cutoff}{the probability level at which to prune links (i.e. cases can not be
linked if the distance or time difference is greater than the this \%ile of the distribution)}

\item{params}{list of parameters to pass to si_fun and dist_fun}
}
\value{
a data.table with the reconstructed tree
}
\description{
Reconstruct transmission trees from a case line list.
}
