% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{check_lineages}
\alias{check_lineages}
\title{Get the number of lineages in each chain}
\usage{
check_lineages(links)
}
\arguments{
\item{links}{}
}
\value{

}
\description{
Get the number of lineages in each chain
}
