% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{check_loops}
\alias{check_loops}
\title{Helper function to check for loops}
\usage{
check_loops(links)
}
\arguments{
\item{links}{either the consensus links or a single tree (with
cols id_case & id_progen)}
}
\value{
a vector of case ids which are part of a loop in the tree
}
\description{
Helper function to check for loops
}
