% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{est_detection}
\alias{est_detection}
\title{Function to get likelihood of a detection probability given observed distances
between case pairs and an underlying distribution}
\usage{
est_detection(
  x,
  p = NULL,
  alpha,
  pdf = c("temporal", "spatial", "empirical", "genetic"),
  fixed_pars = list(),
  par = c(pi = 0.1),
  adj = 1e-06
)
}
\arguments{
\item{x}{distances between case pairs}

\item{p}{vector of probabilities corresponding to expected probabilities
of discrete sequential distances (this is still not ironed out, but only
needed for empirical distributions)}

\item{alpha}{the value at which to cutoff the potential number of cases between two linked cases
see Cori et al. for more details}

\item{pdf}{type of distance, options are "temporal", "spatial", "empirical", or "genetic"}

\item{fixed_pars}{parameters to pass to the d functions (i.e. dempiric, dgamma, dgenetic, dspatial)}

\item{par}{starting value for detection probability}

\item{adj}{constraining distances to be non-zero}
}
\value{

}
\description{
Function to get likelihood of a detection probability given observed distances
between case pairs and an underlying distribution
}
