% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{estimate_detection}
\alias{estimate_detection}
\title{Estimate detection probabilities given observed distances between case pairs}
\usage{
estimate_detection(
  dist_pairs,
  p = NULL,
  pdf = c("temporal", "spatial", "empirical", "genetic"),
  fixed_pars,
  alpha = 0.001
)
}
\arguments{
\item{dist_pairs}{observed distances between linked case pairs (i.e. from
contact tracing data + transmission tree reconstruction)}

\item{p}{vector of probabilities corresponding to expected probabilities
of discrete sequential distances (this is still not ironed out, but only
needed for empirical distributions)}

\item{pdf}{type of distance, options are "temporal", "spatial", "empirical", or "genetic"}

\item{fixed_pars}{parameters to pass to the d functions (i.e. dempiric, dgamma, dgenetic, dspatial)}

\item{alpha}{the value at which to cutoff the potential number of cases between two linked cases
see Cori et al. for more details}
}
\value{
an optim object
}
\description{
Adapted from \url{Cori et al. 2019} and the R package \code{vimes}
}
