% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{find_lineages}
\alias{find_lineages}
\title{Find the links to fix too resolve lineage discrepancies}
\usage{
find_lineages(gr, links, known_progens)
}
\arguments{
\item{gr}{the graph of the links}

\item{links}{links with original chain membership and lineages}

\item{known_progens}{a numeric vector of case ids for which progenitors are known}
}
\value{
a numeric vector of case ids to fix
}
\description{
Find the links to fix too resolve lineage discrepancies
}
