% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{find_lins_to_fix}
\alias{find_lins_to_fix}
\title{Internal function for finding mismatched lineages to fix}
\usage{
find_lins_to_fix(links_consensus, known_progens, selector)
}
\arguments{
\item{links_consensus}{consensus links build within build_consensus_links function}

\item{known_progens}{a numeric vector of case ids for which progenitors are known}
}
\value{
a numeric vector of case ids for which links should be reassigned
}
\description{
Internal function for finding mismatched lineages to fix
}
\keyword{internal}
