% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{find_loops}
\alias{find_loops}
\title{Internal function for finding loops}
\usage{
find_loops(links_consensus)
}
\arguments{
\item{links_consensus}{}
}
\value{

}
\description{
Internal function for finding loops
}
\keyword{internal}
