% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_chains.R
\name{get_chain_membership}
\alias{get_chain_membership}
\title{Get membership (i.e. chain id) for each case}
\usage{
get_chain_membership(gr)
}
\arguments{
\item{gr}{graph of transmission tree}
}
\value{
a data.table with id_case, membership, and t_days (time in days from
the start date, i.e. the earliest case in the transmission tree)
}
\description{
Get membership (i.e. chain id) for each case
}
