% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_chains.R
\name{get_chain_persistence}
\alias{get_chain_persistence}
\title{Get duration of chain persistence in days}
\usage{
get_chain_persistence(chain_membership)
}
\arguments{
\item{chain_membership}{output from \code{get_chain_membership} (a data.table
with the membership of each case and the time in days)}
}
\value{
a data.table with the start and end days (from the origin, i.e. minimum
case date in the transmission tree) and the duration of the chain in days.
}
\description{
Get duration of chain persistence in days
}
