% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_chains.R
\name{get_chain_stats}
\alias{get_chain_stats}
\title{Get chain size & lengths}
\usage{
get_chain_stats(gr)
}
\arguments{
\item{gr}{graph of transmission tree}
}
\value{
a data.table with membership (i.e. chain id), chain size, and chain length
}
\description{
Get chain size & lengths
}
