% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{get_edge_dt}
\alias{get_edge_dt}
\title{Internal function for getting edges between mismatched lineages}
\usage{
get_edge_dt(gr, lins)
}
\arguments{
\item{lins}{}
}
\value{

}
\description{
Internal function for getting edges between mismatched lineages
}
