% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{get_gr}
\alias{get_gr}
\title{Helper function to get graph from links}
\usage{
get_gr(links)
}
\arguments{
\item{links}{consensus links or individual tree}
}
\value{
an igraph object
}
\description{
Helper function to get graph from links
}
