% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_chains.R
\name{get_graph}
\alias{get_graph}
\title{Turn a transmssion tree into a graph object}
\usage{
get_graph(ttree)
}
\arguments{
\item{ttree}{data.table with columns \code{id_progen} (id of case progenitor),
\code{id_case} (id of case), \code{t} (date of case), \code{x_coord}, and \code{y_coord}}
}
\value{
a graph generated by igraph
}
\description{
Turn a transmssion tree into a graph object
}
