% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{get_membership}
\alias{get_membership}
\title{Internal function for checking membership of chains AFTER lineage mismatches
& loops broken}
\usage{
get_membership(links)
}
\arguments{
\item{links}{}
}
\value{

}
\description{
Internal function for checking membership of chains AFTER lineage mismatches
& loops broken
}
