% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{select_progenitor}
\alias{select_progenitor}
\title{Wrapper to select single progenitor in each case}
\usage{
select_progenitor(tree, si_fun, dist_fun, params)
}
\arguments{
\item{tree}{the data.table with possible case pairs to select from}

\item{si_fun}{the function to get the probability of a given temporal
difference between cases in days (i.e. the serial interval)}

\item{dist_fun}{the function to get the probability of a given spatial
difference between cases in meters (i.e. the dispersal kernel)}

\item{params}{list of parameters to pass to si_fun and dist_fun}
}
\value{
a data.table filtered to the selected case-progenitor pair
}
\description{
Wrapper to select single progenitor in each case
}
\keyword{internal}
