% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{si_fun_lnorm}
\alias{si_fun_lnorm}
\title{Example serial interval function for input}
\usage{
si_fun_lnorm(N, params)
}
\arguments{
\item{N}{the number to draw}

\item{params}{a list with the parameters}
}
\value{
a vector of serial intervals
}
\description{
Example serial interval function for input
}
