% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{si_gamma1}
\alias{si_gamma1}
\title{Gamma distribution for probabilities and cutoff for serial interval}
\usage{
si_gamma1(ttree, params, cutoff = NULL)
}
\description{
These are examples for how to write the si_fun and dist_fun functions.
They must take three parameters: ttree, params, and cutoff.
}
\details{
If the cutoff is NULL, it will take the ttree and add a new column
with the probability (using the column t_diff which is generated inside
build_tree). If the cutoff is not NULL then it will return the value at which
you should prune the trees--this should either return a single value of length 1
or a vector of values of length nrow(ttree). See dist_gamma_mixed for an example.
}
