% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{sim_times_pi}
\alias{sim_times_pi}
\title{Simulate times given generation function & pi, for validating detection estimation}
\usage{
sim_times_pi(si_fun, nobs, params, alpha = 0.001, pi)
}
\arguments{
\item{si_fun}{a function for the serial interval with arguments N (the number
to draw) and params (a list of parameters for the function), see si_fun_lnorm for
an example.}

\item{nobs}{the number of observations to simulation}

\item{params}{a list with parameters for the si_fun function to draw serial intervals}

\item{alpha}{probability, the value at which to constrain kappa (i.e. to determine max_kappa
for sim_generations), i.e. the probability of observing this kappa for a given pi is < alpha}

\item{pi}{detection probability (the proportion of cases which are detected)}
}
\value{
a vector of time differences between linked cases
}
\description{
Simulate times given generation function & pi, for validating detection estimation
}
