% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_links.R
\name{build_consensus_links}
\alias{build_consensus_links}
\title{Build consensus links between cases (i.e. most often selected progenitors for a given case)}
\usage{
build_consensus_links(all_links, case_dates)
}
\arguments{
\item{all_links}{output from \code{build_all_links}}

\item{case_dates}{data.table with two columns id_case (id of case) and symptoms_started
(date symptoms started of case without uncertainty).}
}
\value{
a data.table with the case pairs most often linked in transmission tree and their
probability (i.e. prop of times a given case was selected as a progenitor)
}
\description{
Build consensus links between cases (i.e. most often selected progenitors for a given case)
}
