% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{build_known_tree}
\alias{build_known_tree}
\title{Reconstruct known tree (i.e. from tracing data)}
\usage{
build_known_tree(t_dt, t_dt_on, si_fun, dist_fun, params, min_time = 1e-06)
}
\arguments{
\item{t_dt}{made inside build_tree}

\item{t_dt_on}{made inside build_tree}

\item{si_fun}{the function to get the probability of a given temporal
difference between cases in days (i.e. the serial interval)}

\item{dist_fun}{the function to get the probability of a given spatial
difference between cases in meters (i.e. the dispersal kernel)}

\item{params}{list of parameters to pass to si_fun and dist_fun}

\item{min_time}{if uncertainty results in negative time difference between
known case pairs, set it to this value (better way to deal with propagating
uncertainty given known case pairs?)}
}
\value{
a data.table with the known contact tracing tree
}
\description{
Reconstruct known tree (i.e. from tracing data)
}
\keyword{internal}
